batch processing complete

+++ ChangeDetection/processing/process_all_s2.py
@@ -4,7 +4,7 @@ import os
 from osgeo import gdal
 import numpy as np
 import boto3
-
+import botocore
 
 from processing_utils import mask_clouds_and_calc_ndvi, arr_to_gtiff
 
@@ -34,12 +34,16 @@ def calc_ndvi_and_mask_s2_clouds(file):
     # in theory we should be able to use the vector file to mask out clouds, but
     # in practice gdal had complaints about the gml files. will reinvestigate this
     # at a later date
-    cloud_mask_tif = gdal.Rasterize("mask.tif", cloud_mask_band_file, xRes=10, yRes=10, burnValues=1, 
+    try:
+        cloud_mask_tif = gdal.Rasterize("mask.tif", cloud_mask_band_file, xRes=10, yRes=10, burnValues=1, 
                                     noData=np.nan, outputType=gdal.GDT_Byte, outputBounds=[x_min, y_min, x_max, y_max])
-    cloud_mask_tif = None
-    cloud_mask_ds = gdal.Open("mask.tif")
-    cloud_mask = cloud_mask_ds.GetRasterBand(1).ReadAsArray()
-    
+        cloud_mask_tif = None
+        cloud_mask_ds = gdal.Open("mask.tif")
+        cloud_mask = cloud_mask_ds.GetRasterBand(1).ReadAsArray()
+    except AttributeError as e:
+        print(f"Could not read cloud mask file {cloud_mask_band_file}")
+        return None
+
     ndvi_masked = mask_clouds_and_calc_ndvi(red, nir, cloud_mask)
 
     # remove temp cloud mask file to save space
@@ -64,33 +68,32 @@ def main():
     granules = []
     for key in response['Contents']:
             path = os.path.dirname(key['Key'])
-            granule = f"{path}/{os.path.basename(path)}"
+            granule = f"{src_bucket}/{path}/{os.path.basename(path)}"
             try:
                 # check if file already exists
-                s3.Object(dest_bucket, f"{granule}_NDVI_MASKED.TIF").load()
+                s3.head_object(Bucket=dest_bucket, Key=f"{path}_NDVI_MASKED.TIF")
             except botocore.exceptions.ClientError as e:
                 if e.response['Error']['Code'] == "404":
-                    granules.append(f"{src_bucket}/{granule}")
+                    granules.append(granule)
                 else:
                     raise e
 
     granules = set(granules)
-
     for granule in granules:
         prefix = granule.split('/', 1)
         bucket = prefix[0]
         key = prefix[1]
         # vsis3 tells gdal that the file is in an s3 bucket
         result = calc_ndvi_and_mask_s2_clouds(f"/vsis3/{bucket}/{key}")
-        print(f"Generated {result}")
+        if result:
+            print(f"Generated {result}")
         
-        # remove redundant folder name from uploaded file
-        prefix = os.path.dirname(os.path.dirname(key))
-        key = f"{prefix}/{result}"
-        s3 = boto3.client('s3')
-        s3.upload_file(result, dest_bucket, key)
-        print(f"Uploaded {key} to {dest_bucket}")
-        os.remove(result)
+            # remove redundant folder name from uploaded file
+            prefix = os.path.dirname(os.path.dirname(key))
+            key = f"{prefix}/{result}"
+            s3.upload_file(result, dest_bucket, key)
+            print(f"Uploaded {key} to {dest_bucket}")
+            os.remove(result)
     
     print(f"Processed {len(granules)} granules.")
